Long forgotten tradition to play with benchmarks

+++ itoa/itoa.go
@@ -4,7 +4,7 @@ import "errors"
 
 var BufLenErr = errors.New("Buffer is too short")
 
-func NumToBuf(i uint16, buf []byte) (int, error) {
+func NumToBuf(i uint64, buf []byte) (int, error) {
 	digs := 1
 	for lim := uint64(10); lim <= uint64(i); lim *= 10 {
 		digs += 1
